Refactoring `image-generate` to use `image-provision`. Issue #13

+++ test/test_helper.bash
@@ -125,8 +125,13 @@ fixtures() {
 }
 
 cp_fixture() {
-  mkdir -p "$(dirname "$2")"
-  cp "$(fixtures $1).json" $2
+  if [ -e "$2" ]; then
+    mkdir -p "$(dirname "$2")"
+    cp "$(fixtures $1).json" $2
+  else
+    mkdir -p "$2"
+    cp -rf "$(fixtures $1)" "$2"
+  fi
 }
 
 create_file() {
